feat(provider/openai): full implementation for Files API file-ids and urls in Images, PDFs, across Chat and Responses (#7804)
## Background
### New Support

### OpenAI API File ID Support

In the OpenAI API, File IDs from the Files API are supported for the
following use cases that the SDK does not yet support:

#### Relevant Use Cases for FileID

- **PDFs in the Chat API** using FileID
- **PDFs in the Responses API** using FileID
- **Images in the Responses API** usingFileID

### OpenAI API Full URL Support

#### Relevant Use Cases for URL
- **PDFs in the Responses API** using URL (already supported in the
SDK's use of Chat API)


#### Additional Notes

The SDK's workaround for supporting URL inputs for PDFs in the Chat API
was not implemented for the Responses API.
Instead of using the workaround as this PR does, it may be worthwhile to
leverage the Responses API's native support for URL.

## Summary

Added support to reflect what's supported in the API 

The API column represents the OpenAI API
SDK main represents the current state of the AI SDK
SDK PR represents this PR

#### Table 1: Chat API PDF

| Feature | API | SDK Main | SDK PR |
|---------|-----|----------|--------|
| Files API ID | ✅ | ❌ | ✅ |
| URL | ❌ | ✅ | ✅ |
| base64 | ✅ | ✅ | ✅ |

#### Table 2: Chat API Image

| Feature | API | SDK Main | SDK PR |
|---------|-----|----------|--------|
| Files API ID | ❌ | ❌ | ❌ |
| URL | ✅ | ✅ | ✅ |
| base64 | ✅ | ✅ | ✅ |

#### Table 3: Responses PDF

| Feature | API | SDK Main | SDK PR |
|---------|-----|----------|--------|
| Files API ID | ✅ | ❌ | ✅ |
| URL | ✅ | ❌ | ✅ |
| b64 | ✅ | ✅ | ✅ |

#### Table 4: Responses Image

| Feature | API | SDK Main | SDK PR |
|---------|-----|----------|--------|
| Files API ID | ✅ | ❌ | ✅ |
| URL | ✅ | ✅ | ✅ |
| b64 | ✅ | ✅ | ✅ |

## Manual Verification

I ran javascript scripts to query the OpenAI API through fetch in order
to compare expected results against what is in the documentation.

I then made a simple node project using the latest version of the AI SDK
to verify support discrepancies

Finally, I tested the build from my PR to verify the changes worked.

## Future Work

It may be worthwhile to leverage the Responses API's native support for
URL instead of extending the workaround used in Chat to Responses. I
think it was beyond the scope of the PR

## Related Issues
Fixes #7581 
Fixes #7803