fix(codemod): track generateText usage before transform (#7693)
## Background


this is based on https://github.com/vercel/ai/pull/6554#discussion_r2247413087

The current approach replaces all places where result.text,
response.text is used. Its very common to use these functions and
replacing all is incorrect. Also it doesn't transform when the variable
name is different than result or response (and other common names). So
this pr tracks for `generateText` usage, gets the variable name and
replaces them only. This still has issue like if same name variable is
there in the file for other use too, this'll transform that too, but ig
this approach is still better than prev one.

## Summary

I'm tracking usage of `generateText`, getting the variables then
transforming them.

## Verification

I ran the tests, all existing one are passing and added a new test to
make sure it doesnt transform all. The jsshift is returning empty when
no transforms are applied for me, ig its expected behaviour so the
output file is kept empty

## Future Work

This pr still has issue like if same name variable is there in the file
for other use too, this'll transform that too, but ig this approach is
still better than prev one.