feat(provider/google): add labels field to providerOptions (#7928)
## Background

We were trying to break down our Vertex bill but the SDK is filtering
out the labels field.

## Summary

Added the `labels` field to the `providerOptions` Zod schema as
documented
[here](https://cloud.google.com/vertex-ai/generative-ai/docs/multimodal/add-labels-to-api-calls).

## Manual Verification

Verification in progress (takes ~24h for bill to update).

## Future Work

Might make sense to split Google Vertex and Google provider options in
the future as there's quite a bit in here already that seems to be
Vertex-only.