feat(ai): preliminary tool results (#7987)
## Background

Tools often need to send preliminary results, e.g. status information.
We currently recommend using data parts for such use cases, but that
requires creating a writer and related boilerplate code.

In AI SDK 5, we have a `toModelOutput` method, so tool outputs and what
is being sent to LLMs can be easily decoupled. This opens the path for
adding status information and more to tool results, and thus preliminary
tool outputs with the same shape as the final output are possible
without sending irrelevant information to the LLM.

## Summary

* Add support for `AsyncIterable` outputs to tool execute functions
* Add `preliminary` flag to streamed tool results
* Add `preliminary` flag to ui message tool results

## Example

```ts
tool({
  description: "Get the current weather.",
  inputSchema: z.object({ location: z.string() }),
  async *execute({ location }) {
    yield {
      status: "loading" as const,
      text: `Getting weather for ${location}`,
      weather: undefined,
    };

    await new Promise((resolve) => setTimeout(resolve, 3000));
    const temperature = 72 + Math.floor(Math.random() * 21) - 10;

    yield {
      status: "success" as const,
      text: `The weather in ${location} is ${temperature}°F`,
      weather: { temperature },
    };
  },
});
```

## Manual Verification

- [x] streamText console example
- [x] ui example

## Related Issues
Fixes #3063