fix(ai): Added providerOptions to agent.stream and agent.generate (#8034)
## Background

`experimental_Agent` had no way to pass `providerOptions` in the
`stream` or `generate` calls. As a result, there was no way to access
provider options using calls made by `experimental_Agent` in
`LanguageModelV2Middleware`. This has major implications as
`providerOptions` is typically used for things such as enabling
reasoning in many models.

## Summary

Added `providerOptions` to the `stream` and `generate` functions in
`experimental_Agent`, this gets passed down in the resulting
`generateText` or `streamText` calls.

## Manual Verification

I manually tested this and it worked. Essentially, you'd create an agent
that uses a model wrapped in middleware, then pass provider options to
it:

```typescript
agent.stream({
      messages: convertToModelMessages(chatRequest.messages),
      providerOptions
    });
```

## Related Issues

Fixes issue discovered in #7989