fix(ai): call abort callback when stream is aborted during tool execution (#8197)
## Background

Abort signal were only forwarded to tools and fetch, but not processed
for immediate stream abortion. In case of aborting during tool
execution, the request stopped without calling the abort handler or
attaching an abort event.

## Summary

Check the abort signal before each chunk is streamed out and stop the
stream when it is aborted.

## Related Issues

Continues #8106
Fixes #8107

---------

Co-authored-by: istarkov <istarkov@gmail.com>