feat(provider/mistral): export `MistralLanguageModelOptions` type (#8202)
## Background

We are currently not exporting a type for `providerOptions`

## Summary

We have to take into account that provider options can be different
depending on what is used:

- language model
- image model
- embedding

So instead of `MistralProviderOptions` we settled on
`MistralLanguageModelOptions`. We considered
`MistralLanguageModelProviderOptions` but rejected it due to its length.

## Manual Verification


[examples/ai-core/src/generate-text/mistral-provider-options.ts](https://github.com/vercel/ai/blob/a380bdc31176cf81efe5964e75fbc389d134e435/examples/ai-core/src/generate-text/mistral-provider-options.ts#L9-L17)

## Tasks

- [ ] ~~Tests have been added / updated (for bug fixes / features)~~ we
don't currently do test of module exports, we use examples instead
- [x] Documentation has been added / updated (for bug fixes / features)
- [x] A _patch_ changeset for relevant packages has been added (for bug
fixes / features - run `pnpm changeset` in the project root)
- [x] Formatting issues have been fixed (run `pnpm prettier-fix` in the
project root)

## Related Issues

Addresses https://github.com/vercel/ai/pull/8130#discussion_r2289033254