feat(provider/gateway): Add cache pricing fields to model metadata (#8280)
<!--
Welcome to contributing to AI SDK! We're excited to see your changes.

We suggest you read the following contributing guide we've created
before submitting:

https://github.com/vercel/ai/blob/main/CONTRIBUTING.md
-->

## Background

Re:
https://vercel.slack.com/archives/C099VQB75TP/p1755977985474589?thread_ts=1755977765.274169&cid=C099VQB75TP

## Summary

Added cache pricing fields in the model metadata schema

## Manual Verification

Tested with an example script that printed the metadata for each model,
and I verified that cache pricing was printed (not for every model, many
don't support it).

## Tasks

<!--
This task list is intended to help you keep track of what you need to
do.
Feel free to add tasks and remove unnecessary tasks or this section as
needed.

Please check if the PR fulfills the following requirements:
-->

- [X] Tests have been added / updated (for bug fixes / features)
- [ ] Documentation has been added / updated (for bug fixes / features)
- [X] A _patch_ changeset for relevant packages has been added (for bug
fixes / features - run `pnpm changeset` in the project root)
- [X] Formatting issues have been fixed (run `pnpm prettier-fix` in the
project root)
- [X] I have reviewed this pull request (self-review)