docs: update braintrust guide (#8369)
## Background

Improving documentation. When specifying both the environment variable
`OTEL_EXPORTER_OTLP_ENDPOINT` and a `traceExporter: new
BraintrustExporter`, the braintrust exporter alongside it's
configuration is overriden - the exporter used is the auto-configured
one from Vercel.

## Summary

This PR removes the environment variable approach from the Braintrust
documentation to make the preferred method more clear and prevent people
from doing both configuration paths which leads to confusing behavior.
 
## Manual Verification

This docs change aligns with the docs on the Braintrust website
https://www.braintrust.dev/docs/start/frameworks/opentelemetry#nextjs