fix(codemod): handle import alias for maxStep-stopWhen (#8366)
## Background
This is a continuation of #7890 to implement the import alias part. Also
adds warning for `useChat` saying `maxSteps` is removed from client
side.

## Summary
This pr adds import alias handling for `maxSteps` and also adds warning
for using it in `useChat`


## Tasks

- [x] Tests have been added / updated (for bug fixes / features)
- [ ] Documentation has been added / updated (for bug fixes / features)
- [x] A _patch_ changeset for relevant packages has been added (for bug
fixes / features - run `pnpm changeset` in the project root)
- [x] Formatting issues have been fixed (run `pnpm prettier-fix` in the
project root)
- [x] I have reviewed this pull request (self-review)