fix(provider-utils): fix type inference for toModelOutput (#8437)
## Background

The argument of `toModelOutput` was inferred as `unknown`, despite the
whole tool output being inferred correctly.

## Summary

When there is no inferred output, `toModelOutput` should still be
available with `any` as arg, otherwise fully typed. The inference for
toModelOutput needed to be adjusted since it is different from the rest
of the OUTPUT inference (never vs any).