feat(providers/google-vertex): title parameter support for embedding models (#8482)
## Background

The title parameter was not supported in the AI SDK but seems reasonable
for users to be able to benefit from.

## Summary

I added the title parameter to the schema and allow for passing it when
creating embeddings through the vertex provider.
I added a corresponding test.
I added the parameter to the documentation for the provider's embeddings
model.


## Manual Verification
I verified that the title parameter is silently failing. I verified that
it successfully passes after the changes in the PR

## Related Issues

Fixes #8040