RFC: Pass addToolResult into onToolCall callback
This commit adds two new additional arguments to `onToolCall` the
`addToolResult` function bound to the currently active response and
the `AbortSignal` associated with the active response.

This fixes an issue with the current tool calling flow and cancellation
where if a tool is evaluated and `addToolResult` is called after
`stop()` is called to abort the response then `addToolResult` will
update messages of the new stream if any or triggering a new
`sendMessage()` call if not.

This commit leaves the existing `addToolResult` unbound on the public
interface because I think there are valid use cases for updating a
tool call outside of the current request cycle.