fix(ai): update docs for NoSpeechGeneratedError and export it (#8512)
## Background

The documention mentions `NoAudioGeneratedError`, but this type does not
exist in the codebase; instead, there is a `NoSpeechGeneratedError`.
Additionally, I noticed that `NoSpeechGeneratedError` is not exported
externally.

## Summary

The document has been corrected, and this type of error has been
exported.

## Manual Verification

Modifying the text and export type should not cause any issues.

## Tasks

- [ ] Tests have been added / updated (for bug fixes / features)
- [x] Documentation has been added / updated (for bug fixes / features)
- [x] A _patch_ changeset for relevant packages has been added (for bug
fixes / features - run `pnpm changeset` in the project root)
- [x] Formatting issues have been fixed (run `pnpm prettier-fix` in the
project root)
- [x] I have reviewed this pull request (self-review)

## Related Issues

Close #8466

---------

Co-authored-by: Gregor Martynus <39992+gr2m@users.noreply.github.com>