Update next example to show case resumable and abort-able streams (#8613)
## Background

Users run into issues with the combination and aborting streams and
resumable streams. Specifically the problem that users abort a stream in
the front-end without the stream being aborted from the provider, ending
up using tokens for a stream that is no longer needed.

## Summary

Use `next` example to showcase how to abort a stream from a provider by
adding an additional route and checking the chat store in `streamText()`

The pull request also fixes a bug which canceled the stream from the
provider on navigation. Now the same stream continues on page reload or
when opened on a another device.

## Manual Verification

Ran the example locally

## Tasks

- [x] store `canceledAt` as property on chat if stream was stopped

## Checklist
- [ ] n/a ~~Tests have been added / updated (for bug fixes / features)~~
- [x] Documentation has been added / updated (for bug fixes / features)
- [ ] n/a ~~A _patch_ changeset for relevant packages has been added
(for bug fixes / features - run `pnpm changeset` in the project root)~~
- [x] Formatting issues have been fixed (run `pnpm prettier-fix` in the
project root)
- [x] I have reviewed this pull request (self-review)

## Future Work

Discuss with @nicoalbanese about turning this example into a
cookbook/guide as the question about aborting streams comes up a lot.

---------

Co-authored-by: Gregor Martynus <39992+gr2m@users.noreply.github.com>