run prettier on automerge (#9056)
I have see our team asking contributors to run `prettier-fix` in order
to make the CI pass. We can automate that. With this workflow, each time
we activate the auto-merge feature, prettier is run, and if there are
any changes, they are pushed back to the pull request, which should
resolve any CI error due to the prettier check, and save all of us some
time.