fix(ai): back version support v2 (ProviderV2) (#9151)
## Background

We recently upgraded specs to be V3 instead of V2. But for community
developers still using V2 specs, there needed to be an adapter. This fix
is for ProviderV2 to be supported. See [this
comment](https://github.com/vercel/ai/pull/8893#pullrequestreview-3266032968)
for more information

## Summary

Added `ProviderV2` to be accepted as part of the fallbackProvider but
still return `ProviderV3`

## Manual Verification

New tests added

- [x] Tests have been added / updated (for bug fixes / features)
- [ ] Documentation has been added / updated (for bug fixes / features)
- [x] A _patch_ changeset for relevant packages has been added (for bug
fixes / features - run `pnpm changeset` in the project root)
- [x] Formatting issues have been fixed (run `pnpm prettier-fix` in the
project root)

## Future Work

May need to be addressed for `ImageModel`, `TranscriptionModel`, and
`SpeechModel`

## Related Issues

Adresses [this
comment](https://github.com/vercel/ai/pull/8893#pullrequestreview-3266032968)
and fixes #8909