fix(provider/openai): add providerExecuted flag to tool start chunks (#9181)
## Background

When using OpenAI's web search tool, I noticed that `providerExecuted`
field is unset on `tool-input-start` chunks.

`tool-input-start` chunks have a `providerExecuted` field which should
represent whether a tool is executed by the provider or not. This field
is correctly set to true on Anthropic's provider executed tools (on
`tool-input-start` chunks) but that is not the case with OpenAI.

## Summary

Add providerExecuted flag to tool start chunks for web search, etc.

## Related Issues

Fixes  https://github.com/vercel/ai/issues/9180

---------

Co-authored-by: Lars Grammel <lars.grammel@gmail.com>