docs: refactor agents section around stabilised agent class (#9245)
## Background

Now that the `Agent` class is marked as stable (in beta), Agents section
should feature it exclusively.

## Summary

Removed usage of core functions throughout section (except for
workflows). Removed Agent class page and instead merged it into
Overview.

## Checklist


- [ ] Tests have been added / updated (for bug fixes / features)
- [ ] Documentation has been added / updated (for bug fixes / features)
- [ ] A _patch_ changeset for relevant packages has been added (for bug
fixes / features - run `pnpm changeset` in the project root)
- [x] Formatting issues have been fixed (run `pnpm prettier-fix` in the
project root)
- [ ] I have reviewed this pull request (self-review)

## Future Work

<!--
Feel free to mention things not covered by this PR that can be done in
future PRs.
Remove the section if it's not needed.
 -->

## Related Issues

<!--
List related issues here, e.g. "Fixes #1234".
Remove the section if it's not needed.
-->