feat(ai): rootSpan flag for telemetry options
Adds a new flag `rootSpan` to `experimental_telemetry` that instructs to
create a root span (detached from any parent) during the execution. Defaults
to false keeping the current behavior.

The reason to propose this is that currently I'm having a hard time
getting telemetry to show up correctly in langfuse while using effect.

Because the stream returned by stream operations exists outside the
lifecycle of the calling function, traces get reported out of order, and
ultimately certain attributes that should be made available to the root
span never make it.

So forcing the ai traces to be isolated isi the only way I've been able to
sort of solve it.