feat (provider/anthropic): Add support for pdf files in tool result. (#9148)
## Background
Anthropic API supports handling of pdf files in the tool results which
was not supported by AI SDK currently, so added the support for that.

## Summary
Added support for handling pdf files in tool response in anthropic
provider.

## Manual Verification
Wrote a eg with a tool returning pdf file in tool result as base64
string.

## Checklist

- [x] Tests have been added / updated (for bug fixes / features)
- [ ] Documentation has been added / updated (for bug fixes / features)
- [x] A _patch_ changeset for relevant packages has been added (for bug
fixes / features - run `pnpm changeset` in the project root)
- [x] Formatting issues have been fixed (run `pnpm prettier-fix` in the
project root)
- [x] I have reviewed this pull request (self-review)


## Related Issues
Fixes #8986

---------

Co-authored-by: Ankit Varshney <avtheking@Ankits-MacBook-Air.local>