ci(bundle-size): initial version (#9411)
This is just a trivial check to start. It generates two bundles 

1. with condition "node"
2. with condition "browser"

I was actually surprised to see that the browser build is _bigger_ than
the node build. It seems mostly be caused by the open telemetry
dependency.

and it generates meta files that can be uploaded to
https://esbuild.github.io/analyze/ for analysis. They are uploaded as
artifacts to the CI check run.

The max size is currently configured as repository variable:
https://github.com/vercel/ai/settings/variables/actions/BUNDLE_SIZE_LIMIT_KB

---------

Co-authored-by: vercel-ai-sdk[bot] <225926702+vercel-ai-sdk[bot]@users.noreply.github.com>