feat(ui): add isDataUIPart helper (#9580)
## Background

Often, I need to tell if a part is some kind of data part. For example,
if I want to have a generic renderer for all data parts. I could write
`part.type.startsWith('data-')`, but it's much more convenient to import
a helper.

## Summary

Added the function `isDataUiPart` as an analog to `isToolUIPart`.

## Manual Verification

I imported and used this function in my application with `pnpm i <ai
package checkout path>` and validated it works as expected.

---------

Co-authored-by: Lars Grammel <lars.grammel@gmail.com>