docs: make examples deployable again (#9689)
## Background

Deploying examples used to work but was broken via #5864

We have to replace workspace:* with latest versions of the the
respective packages.

## Summary

Replaced all `workspace:*` version values in package.json files with the
result of `npm info <pkg>@beta version`

## Manual Verification

I manually run some of the examples locally.

## Checklist

- [ ] n/a ~~Tests have been added / updated (for bug fixes / features)~~
- [x] Documentation has been added / updated (for bug fixes / features)
- [ ] n/a ~~A _patch_ changeset for relevant packages has been added
(for bug fixes / features - run `pnpm changeset` in the project root)~~
- [x] I have reviewed this pull request (self-review)

## Future Work

Consider backporting, which will need to be done manually.

I used this prompt in cursor which worked for about 90% 🤪

> We need to replace set latest version of each package that is
currently configurde with "workspace:*" in the examples/* folder, except
for examples/ai-core. In order to get the latest version of a package,
run `npm info <pgk>@beta version`, e.g. for `@ai-sdk/angular` run `npm
info @ai-sdk/angular@beta version` and use the result as value for the
version in the package.json file

For the backport, remove the `@beta` from the prompt

## Related Issues

closes #8910