feat(mcp): tool title (#9200)
## Background
MCP supports for human readable title for the tool which is currently
not supported by ai sdk ToolUIPart.

## Summary

Add a way to be able to use the mcp tool title in ui.

## Manual Verification

Wrote an next js eg with mcp server and tools with title and using that
title in the ui as display name for the tool.

## Checklist

- [x] Tests have been added / updated (for bug fixes / features)
- [ ] Documentation has been added / updated (for bug fixes / features)
- [x] A _patch_ changeset for relevant packages has been added (for bug
fixes / features - run `pnpm changeset` in the project root)
- [x] Formatting issues have been fixed (run `pnpm prettier-fix` in the
project root)
- [x] I have reviewed this pull request (self-review)

## Related Issues
Fixes #9163

---------

Co-authored-by: Ankit Varshney <avtheking@Ankits-MacBook-Air.local>
Co-authored-by: Ankit Varshney <varankit8754@gmail.com>
Co-authored-by: vercel-ai-sdk[bot] <225926702+vercel-ai-sdk[bot]@users.noreply.github.com>
Co-authored-by: Lars Grammel <lars.grammel@gmail.com>
Co-authored-by: Gregor Martynus <39992+gr2m@users.noreply.github.com>
Co-authored-by: Aayush Kapoor <83492835+aayush-kapoor@users.noreply.github.com>